Chore: #67 - RoutingSection 컴포넌트 분리
